requirements.txt
fastapi
uvicorn
requests
google-genai
python-dotenv
apscheduler
python-slugify
pydantic-settings